[harness] Support multiple W3C test suites
If the specified report data file contains a URL in a comment line, it will
be used as the URL prefix for each test.
Note unless -x option is specified, 'localhost' is used as the host instead
of the specified one such as 'test.csswg.org'.

(src/harness.cpp): Refine